Update/optimize ibuffer sorting/hiding
There's no need to set the default sorting mode via a hook, ibuffer
exports a variable for it. Secondly, prefer `ibuffer-redisplay` over
`ibuffer-update` - the former just visually updates the buffer, whilst
the latter also checks for new buffers and changes. In the hide hook
we're only interested in changing visuals.