Don't allow Ctrl-Alt-Backspace to kill Xorg
I never use it. I can't remember the last time Xorg and/or my video
drivers crapped out on me... and if even if they did, killing Xorg
didn't fix it most of the time, a power cycle was required. But I've hit
the combination by accident a few times too many, so out it goes.